Adding scoring

+++ Maths/AdditionQuestion.m
@@ -14,9 +14,9 @@
 {
     self = [super init];
     if (self) {
-        long randomNumber1 = arc4random_uniform(90)+10;
-        long randomNumber2 = arc4random_uniform(90)+10;
-        long answer = randomNumber1 + randomNumber2;
+        NSInteger randomNumber1 = arc4random_uniform(90)+10;
+        NSInteger randomNumber2 = arc4random_uniform(90)+10;
+        NSInteger answer = randomNumber1 + randomNumber2;
         
         NSString * combinedNumbers = [NSString stringWithFormat:@"%ld + %ld\n", randomNumber1, randomNumber2];
         _question = combinedNumbers;
